+++ lib/include/arch_avx512_type2.h
@@ -54,14 +54,6 @@ aes_cntr_192_submit_vaes_avx512(IMB_JOB *job);
 void
 aes_cntr_256_submit_vaes_avx512(IMB_JOB *job);
 
-/* AES-CTR-BITLEN */
-void
-aes_cntr_bit_128_submit_vaes_avx512(IMB_JOB *job);
-void
-aes_cntr_bit_192_submit_vaes_avx512(IMB_JOB *job);
-void
-aes_cntr_bit_256_submit_vaes_avx512(IMB_JOB *job);
-
 /* AES-ECB */
 void
 aes_ecb_enc_256_vaes_avx512(const void *in, const void *keys, void *out, uint64_t len_bytes);