+++ lib/include/arch_sse_type1.h
@@ -77,17 +77,6 @@ void
 aes_cntr_128_sse(const void *in, const void *IV, const void *keys, void *out, uint64_t len_bytes,
                  uint64_t IV_len);
 
-/* AES-CTR-BITLEN */
-void
-aes_cntr_bit_256_sse(const void *in, const void *IV, const void *keys, void *out, uint64_t len_bits,
-                     uint64_t IV_len);
-void
-aes_cntr_bit_192_sse(const void *in, const void *IV, const void *keys, void *out, uint64_t len_bits,
-                     uint64_t IV_len);
-void
-aes_cntr_bit_128_sse(const void *in, const void *IV, const void *keys, void *out, uint64_t len_bits,
-                     uint64_t IV_len);
-
 /* AES-CCM */
 IMB_JOB *
 aes_cntr_ccm_128_sse(IMB_JOB *job);